用户权限管理

+++ public/javascripts/movie.js
@@ -1,20 +1,23 @@
 
 $(function(){
 	$('.del').click(function(e){
-		var target = $(e.target)
-		var id = target.data('id');
-		var tr = $(this).parents('tr');
-		console.log(id)
-		$.ajax({
-			type: 'delete',
-			url: '/admin/movie?id='+id
-		})
-		.done(function(res){
-			if(res.status === 1){
-				if(tr.length>0){
-					tr.remove()
-				}
-			}
-		})
+		if (confirm("此操作不是恢复，是否删除")) {
+				var target = $(e.target)
+				var id = target.data('id');
+				var tr = $(this).parents('tr');
+				$.ajax({
+					type: 'delete',
+					url: '/movie/list?id=' + id
+				})
+				.done(function(res){
+					if(res.status === 1){
+						if(tr.length > 0){
+							tr.remove()
+						}
+					}
+				})
+		}else {
+			return false
+		}
 	})
 })
